requirements.txt
matplotlib==3.10.3
numpy==2.3.1
openai==1.92.2
pandas==2.3.0
python-dotenv==1.1.1
scikit_learn==1.7.0
torch==2.7.1
transformers==4.52.4
